pyproject.toml
[tool.poetry]
name = "stats"
version = "0.1.0"
description = ""
authors = ["patrick <chodowski.patrick@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
pandas = "^1.3.0"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"
bokeh = "^2.3.3"
google-cloud-bigquery = "^2.22.1"
PyQt5 = "5.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
